requirements.txt
aiohttp
asyncpg
python-dotenv
bcrypt
PyJWT